Moving ContractResultsSection.vue from /contracts to /contract.
