+++ src/components/contract/ContractResultsSection.vue
