+++ src/pages/NftDetails.vue
@@ -192,7 +192,7 @@ import NotificationBanner from "@/components/NotificationBanner.vue"
 import PlayPauseButton from "@/components/PlayPauseButton.vue"
 import AccountLink from "@/components/values/link/AccountLink.vue"
 import {NftBySerialCache} from "@/utils/cache/NftBySerialCache"
-import ContractResultsSection from "@/components/contracts/ContractResultsSection.vue"
+import ContractResultsSection from "@/components/contract/ContractResultsSection.vue"
 import NftTransactionTable from "@/components/transaction/NftTransactionTable.vue"
 import {NftTransactionTableController} from "@/components/transaction/NftTransactionTableController"
 import TransactionFilterSelect from "@/components/transaction/TransactionFilterSelect.vue"
